+++ CHANGELOG.md
@@ -4,10 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## - 2021-07-22
-
 ### Changed
 - tokens linked list became tokens **double** linked list
 - new tokend types: `T_PIPE` `T_REDIRECT` `T_HERE_DOC` `T_FILE`
+
+## - 2021-07-21
+### Added
+- Added `cd` builtin function.
 
 ## - 2021-07-18
 ### Added
@@ -15,10 +18,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Added the `variadic_strjoin()` function, which permits concatenation of an variable number of strings.
 - Added Macros with escaped colors for readline library.
 - Added `unset()` builtin function.
+- Added	`pwd()` function, that print working directory.
 
 ### Changed
 - Refactor `create_prompt()` function using `variadic_strjoin()`.
-- Added	`pwd()` function, that print working directory.
 
 ## - 2021-07-17
 ### Added